split plugins tab
