+++ src/settingsdialog.cpp
@@ -99,7 +99,6 @@ void SettingsDialog::accept()
     return;
   }
 
-  storeSettings(ui->pluginsList->currentItem());
   TutorableDialog::accept();
 }
 
@@ -147,57 +146,6 @@ void SettingsDialog::on_bsaDateBtn_clicked()
 
   Helper::backdateBSAs(qApp->applicationDirPath().toStdWString(),
                        dir.absolutePath().toStdWString());
-}
-
-void SettingsDialog::storeSettings(QListWidgetItem *pluginItem)
-{
-  if (pluginItem != nullptr) {
-    QVariantMap settings = pluginItem->data(Qt::UserRole + 1).toMap();
-
-    for (int i = 0; i < ui->pluginSettingsList->topLevelItemCount(); ++i) {
-      const QTreeWidgetItem *item = ui->pluginSettingsList->topLevelItem(i);
-      settings[item->text(0)] = item->data(1, Qt::DisplayRole);
-    }
-
-    pluginItem->setData(Qt::UserRole + 1, settings);
-  }
-}
-
-void SettingsDialog::on_pluginsList_currentItemChanged(QListWidgetItem *current, QListWidgetItem *previous)
-{
-  storeSettings(previous);
-
-  ui->pluginSettingsList->clear();
-  IPlugin *plugin = static_cast<IPlugin*>(current->data(Qt::UserRole).value<void*>());
-  ui->authorLabel->setText(plugin->author());
-  ui->versionLabel->setText(plugin->version().canonicalString());
-  ui->descriptionLabel->setText(plugin->description());
-
-  QVariantMap settings = current->data(Qt::UserRole + 1).toMap();
-  QVariantMap descriptions = current->data(Qt::UserRole + 2).toMap();
-  ui->pluginSettingsList->setEnabled(settings.count() != 0);
-  for (auto iter = settings.begin(); iter != settings.end(); ++iter) {
-    QTreeWidgetItem *newItem = new QTreeWidgetItem(QStringList(iter.key()));
-    QVariant value = *iter;
-    QString description;
-    {
-      auto descriptionIter = descriptions.find(iter.key());
-      if (descriptionIter != descriptions.end()) {
-        description = descriptionIter->toString();
-      }
-    }
-
-    ui->pluginSettingsList->setItemDelegateForColumn(0, new NoEditDelegate());
-    newItem->setData(1, Qt::DisplayRole, value);
-    newItem->setData(1, Qt::EditRole, value);
-    newItem->setToolTip(1, description);
-
-    newItem->setFlags(newItem->flags() | Qt::ItemIsEditable);
-    ui->pluginSettingsList->addTopLevelItem(newItem);
-  }
-
-  ui->pluginSettingsList->resizeColumnToContents(0);
-  ui->pluginSettingsList->resizeColumnToContents(1);
 }
 
 void SettingsDialog::deleteBlacklistItem()
